+++ absl/CHANGELOG.md
@@ -13,7 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
     functions.
 *   (testing) flagsaver overrides can now be specified in terms of FlagHolder.
 *   (testing) `parameterized.product`: Allows testing a method over cartesian
-    product of parameters values.
+    product of parameters values, specified as a sequences of values for each
+    parameter or as kwargs-like dicts of parameter values.
 
 ## 0.11.0 (2020-10-27)